+++ src/modules/modules.ts
@@ -6,8 +6,8 @@ import {
   getUpdateSchema,
   generateEnumTypeForGraphql,
   generateGenerateGraphQLCrud,
-  generateRowsFieldNameForModuleName,
-  generateRowFieldNameForModuleName,
+  convertWordIntoPlural,
+  convertWordIntoSingular,
   getOrderSchema,
 } from "./modulesHelpers"
 import { getMysqlTableInfo } from "../database/mysql/getTableInfo"
@@ -42,7 +42,7 @@ export const withModule = (moduleProps: WithModuleProps) => {
     let graphqlSchema = [`type ${moduleProps.name}Module {`]
     let listSchema = ""
     let filterSchema = [
-      `input ${generateRowFieldNameForModuleName(
+      `input ${convertWordIntoSingular(
         moduleProps.name
       )}_filter_input {`,
     ]
@@ -102,7 +102,7 @@ export const withModule = (moduleProps: WithModuleProps) => {
       currentModuleRelationships.push(relationshipInfo)
       store.graphql.graphqlKeys.push(camelize(params.module))
       filterSchema.push(
-        `${camelize(params.graphqlKey)}: ${generateRowFieldNameForModuleName(
+        `${camelize(params.graphqlKey)}: ${convertWordIntoSingular(
           params.module
         )}_filter_input`
       )
@@ -122,13 +122,13 @@ export const withModule = (moduleProps: WithModuleProps) => {
       currentModuleRelationships.push(relationshipInfo)
       store.graphql.graphqlKeys.push(camelize(params.module))
       filterSchema.push(
-        `${camelize(params.graphqlKey)}: ${generateRowFieldNameForModuleName(
+        `${camelize(params.graphqlKey)}: ${convertWordIntoSingular(
           params.module
         )}_filter_input`
       )
     }
     const belongsToMany = (params: RelationParams) => {
-      let field_name = generateRowFieldNameForModuleName(params.module)
+      let field_name = convertWordIntoSingular(params.module)
       graphqlSchema.push(
         `${params.graphqlKey}(offset: Int, limit: Int, where: ${field_name}_filter_input, order: ${field_name}_order_input): [${params.module}Module]`
       )
@@ -145,13 +145,13 @@ export const withModule = (moduleProps: WithModuleProps) => {
       currentModuleRelationships.push(relationshipInfo)
       store.graphql.graphqlKeys.push(camelize(params.module))
       filterSchema.push(
-        `${camelize(params.graphqlKey)}: ${generateRowFieldNameForModuleName(
+        `${camelize(params.graphqlKey)}: ${convertWordIntoSingular(
           params.module
         )}_filter_input`
       )
     }
     const hasMany = (params: RelationParams) => {
-      let field_name = generateRowFieldNameForModuleName(params.module)
+      let field_name = convertWordIntoSingular(params.module)
       graphqlSchema.push(
         `${params.graphqlKey}(offset: Int, limit: Int, where: ${field_name}_filter_input, order: ${field_name}_order_input): [${params.module}Module]`
       )
@@ -168,7 +168,7 @@ export const withModule = (moduleProps: WithModuleProps) => {
       store.database.relationships.push(relationshipInfo)
       store.graphql.graphqlKeys.push(camelize(params.module))
       filterSchema.push(
-        `${camelize(params.graphqlKey)}: ${generateRowFieldNameForModuleName(
+        `${camelize(params.graphqlKey)}: ${convertWordIntoSingular(
           params.module
         )}_filter_input`
       )
@@ -199,7 +199,7 @@ export const withModule = (moduleProps: WithModuleProps) => {
       const connection = app.database[moduleProps.database]
       // info
       const tableInfo = await getMysqlTableInfo(
-        moduleProps,
+        moduleProps.table,
         connection.instance
       )